feat 複数のグループが存在する場合でも対応可能にする
