+++ src/components/Header.vue
@@ -18,6 +18,7 @@ const state = reactive({
   title: props.defaultTitle,
   inputTitle: props.defaultTitle,
   logoName: props.defaultLogoName,
+  loading: false,
 });
 const $PlayerStore = usePlayerStore();
 const $ProfileStore = useProfileStore();
@@ -27,10 +28,12 @@ const getImageUrl = (logoName: string) => {
   return new URL(`/src/assets/${logoName}.svg`, import.meta.url).href
 }
 const save = async () : Promise<void> => {
+  state.loading = true;
   try {
     await window.electronAPI.delete();
     for (const player of $PlayerStore.players) {
-      await window.electronAPI.save(player.toProfileDto(), player.toMatchDtoList());
+      console.log(player.toProfileDto());
+      await window.electronAPI.save(player.profile.group_id, player.toProfileDto(), player.toMatchDtoList());
     }
     const titleInfoDto: TitleInfoDto = {logo_name: state.logoName, title: state.title};
     await window.electronAPI.saveTitleInfo(titleInfoDto);
@@ -38,6 +41,8 @@ const save = async () : Promise<void> => {
     console.error('DB error:', err);
     $toast.error("保存に失敗しました", { position: "top" });
     return;
+  } finally {
+    state.loading = false;
   }
   $toast.success("保存に成功しました!", { position: "top" });
 }
@@ -57,73 +62,86 @@ const reset = (): void => {
 </script>
 
 <template>
-  <header>
-    <v-tooltip location="bottom" text="クリックして編集する">
-      <template v-slot:activator="{ props }">
-        <img v-bind="props" @click="state.logoDialog = true" alt="logo" class="logo" :src="getImageUrl(state.logoName)" />
-        <v-dialog v-model="state.logoDialog" width="auto">
-          <v-card width="650">
-            <v-card-title class="daialog-title">ロゴの選択</v-card-title>
-            <v-sheet class="px-3">
-              <div class="image-container">
-                <img alt="logo" class="pa-3" src="@/assets/igo.svg" @click="state.logoDialog = false; state.logoName='igo'" />
-                <img alt="logo" class="pa-3" src="@/assets/shogi.svg" @click="state.logoDialog = false; state.logoName='shogi'" />
-                <img alt="logo" class="pa-3" src="@/assets/chess.svg" @click="state.logoDialog = false; state.logoName='chess'" />
-                <img alt="logo" class="pa-3" src="@/assets/trophy.svg" @click="state.logoDialog = false; state.logoName='trophy'" />
-              </div>
-            </v-sheet>
-          </v-card>
-        </v-dialog>
-      </template>
-    </v-tooltip>
-    <h1 class="title">
-      {{ state.title }}
-      <sup @click="state.titleDialog = true">
-        <v-tooltip location="top"
-          text="クリックして編集する">
-          <template v-slot:activator="{ props }">
-            <v-icon v-bind="props" opacity="40%" size="x-small" icon="mdi-pencil-outline" class="title-edit-icon"></v-icon>
-          </template>
-        </v-tooltip>
-      </sup>
-    </h1>
-    <v-dialog v-model="state.titleDialog" width="auto">
-      <v-card width="500">
-        <v-card-title class="title-daialog">タイトルの入力</v-card-title>
-        <v-sheet class="px-3">
-          <v-text-field v-model="state.inputTitle" variant="underlined" density="compact" maxlength="30"
-            :rules="[validatorUtil.checkTitleLength]">
-          </v-text-field>
-        </v-sheet>
-        <v-divider></v-divider>
-        <template v-slot:actions>
-          <v-btn class="bg-grey-lighten-2" text="キャンセル" @click="state.titleDialog = false; state.inputTitle = state.title;"></v-btn>
-          <v-btn class="bg-green-darken-1 text-white" text="登録する" @click="state.titleDialog = false; state.title = state.inputTitle;"></v-btn>
+  <div class="header-wrapper">
+    <header>
+      <v-tooltip location="bottom" text="クリックして編集する">
+        <template v-slot:activator="{ props }">
+          <img v-bind="props" @click="state.logoDialog = true" alt="logo" class="logo" :src="getImageUrl(state.logoName)" />
+          <v-dialog v-model="state.logoDialog" width="auto">
+            <v-card width="650">
+              <v-card-title class="daialog-title">ロゴの選択</v-card-title>
+              <v-sheet class="px-3">
+                <div class="image-container">
+                  <img alt="logo" class="pa-3" src="@/assets/igo.svg" @click="state.logoDialog = false; state.logoName='igo'" />
+                  <img alt="logo" class="pa-3" src="@/assets/shogi.svg" @click="state.logoDialog = false; state.logoName='shogi'" />
+                  <img alt="logo" class="pa-3" src="@/assets/chess.svg" @click="state.logoDialog = false; state.logoName='chess'" />
+                  <img alt="logo" class="pa-3" src="@/assets/trophy.svg" @click="state.logoDialog = false; state.logoName='trophy'" />
+                </div>
+              </v-sheet>
+            </v-card>
+          </v-dialog>
         </template>
-      </v-card>
-    </v-dialog>
-    <v-tooltip location="top" text="リセット">
-      <template v-slot:activator="{ props }">
-        <v-btn v-bind="props" icon="mdi-reload" class="reset-button text-white text-body-1" @click="reset();"></v-btn>
-      </template>
-    </v-tooltip>
-    <v-tooltip location="top" text="保存">
-      <template v-slot:activator="{ props }">
-        <v-btn v-bind="props" icon="mdi-content-save" class="save-button text-white text-body-1" @click="save();"></v-btn>
-      </template>
-    </v-tooltip>
-    <v-tooltip location="top" text="印刷">
-      <template v-slot:activator="{ props }">
-        <v-btn v-bind="props" icon="mdi-printer-outline" class="print-button text-white text-body-1" @click="print();"></v-btn>
-      </template>
-    </v-tooltip>
-  </header>
+      </v-tooltip>
+      <h1 class="title">
+        {{ state.title }}
+        <sup @click="state.titleDialog = true">
+          <v-tooltip location="top"
+            text="クリックして編集する">
+            <template v-slot:activator="{ props }">
+              <v-icon v-bind="props" opacity="40%" size="x-small" icon="mdi-pencil-outline" class="title-edit-icon"></v-icon>
+            </template>
+          </v-tooltip>
+        </sup>
+      </h1>
+      <v-dialog v-model="state.titleDialog" width="auto">
+        <v-card width="500">
+          <v-card-title class="title-daialog">タイトルの入力</v-card-title>
+          <v-sheet class="px-3">
+            <v-text-field v-model="state.inputTitle" variant="underlined" density="compact" maxlength="30"
+              :rules="[validatorUtil.checkTitleLength]">
+            </v-text-field>
+          </v-sheet>
+          <v-divider></v-divider>
+          <template v-slot:actions>
+            <v-btn class="bg-grey-lighten-2" text="キャンセル" @click="state.titleDialog = false; state.inputTitle = state.title;"></v-btn>
+            <v-btn class="bg-green-darken-1 text-white" text="登録する" @click="state.titleDialog = false; state.title = state.inputTitle;"></v-btn>
+          </template>
+        </v-card>
+      </v-dialog>
+      <v-tooltip location="top" text="リセット">
+        <template v-slot:activator="{ props }">
+          <v-btn v-bind="props" icon="mdi-reload" class="reset-button text-white text-body-1" @click="reset();"></v-btn>
+        </template>
+      </v-tooltip>
+      <v-tooltip location="top" text="保存">
+        <template v-slot:activator="{ props }">
+          <v-btn v-bind="props" icon="mdi-content-save" class="save-button text-white text-body-1" @click="save();"></v-btn>
+        </template>
+      </v-tooltip>
+      <v-tooltip location="top" text="印刷">
+        <template v-slot:activator="{ props }">
+          <v-btn v-bind="props" icon="mdi-printer-outline" class="print-button text-white text-body-1" @click="print();"></v-btn>
+        </template>
+      </v-tooltip>
+    </header>
+    <v-progress-linear
+      v-if="state.loading"
+      :indeterminate="state.loading"
+      color='primary'
+      height='10'
+    ></v-progress-linear>
+  </div>
 </template>
 
 <style scoped>
 img {
   /* 参考URL：https://angel-rs.github.io/css-color-filter-generator */
   filter: brightness(0) saturate(100%) invert(96%) sepia(8%) saturate(300%) hue-rotate(42deg) brightness(97%) contrast(89%);
+}
+.header-wrapper {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
 }
 header {
   /* 横並び */
